+++ spl/src/governance.rs
@@ -60,7 +60,7 @@ macro_rules! vote_weight_record {
 
         #[cfg(feature = "idl-build")]
         impl anchor_lang::Discriminator for VoterWeightRecord {
-            const DISCRIMINATOR: [u8; 8] = [0; 8];
+            const DISCRIMINATOR: &[u8] = &[];
         }
     };
 }